api/requirements.txt
Flask
numpy
mysql-connector